msm: vidc: Reset Venus firmware cookie during failure
When there is an error while loading Venus firmware, the
cookie tracking the firmware is not reset to NULL. Due
to this, there is no further attempt to download the
firmware for any subsequent video session. Unless the
device is rebooted, firmware fails to download and hence
video session continue to fail.
Reset the cookie to NULL to attempt firmware download
during subsequent video session.

Signed-off-by: Vikash Garodia <vgarodia@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>