msm: vidc: cancel power thread if not scheduled
During power collapse there is a race condition where power thread is
trying to collapse venus and other thread is trying to map a buffer.
This could lead to a iommu failure and crash as during power collapse
we detach iommu, and mapping buffers need iommu to be attached. This
change cancels power collapse thread if not scheduled and if executing,
then will wait for the completion of the thread and then we attach
iommu before calling iommu map.

Signed-off-by: Ashray Kulkarni <ashrayk@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>