arm/arm64: irq: drop "IRQx no longer affine to CPUn" print to pr_debug()
On devices where hotplug is used for power savings or thermal
mitigation, it's not uncommon for hotplug events to be relatively
frequent and their latency to matter.

If a serial console is enabled, the "IRQ no longer affine to CPU"
informational prints can be one of the most significant contributors
to hotplug latency. For example, on msm8994, 10 lines of this format
are normally printed on each hotplug. At 115200/8/N/1 baud, that
printing takes a minimum of 42.5ms.

On production devices where a serial console is not enabled, the time
takes is negligible (a few microseconds at most), but the prints do
still clutter the logs and are worth silencing since they are
typically of no concern.

Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>